% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_import.R, R/library_import.R
\name{getMSP}
\alias{getMSP}
\title{getMSP}
\usage{
getMSP(x)

getMSP(x)
}
\arguments{
\item{file}{msp file to read}
}
\value{
list of library entries

a list of list
}
\description{
Read library MSP file

Read lines of msp file
}
