% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCDF_SC.R
\name{readCDF2}
\alias{readCDF2}
\title{readCDF2}
\usage{
readCDF2(filePath, modulationTime, mzRange, ncore = NULL)
}
\arguments{
\item{modulationTime}{modulation time used in the data.}

\item{mzRange}{range values used in the m/z data acquisition.}

\item{ncore}{number of parallel processes to start. Defaults to number of available cores - 1.}

\item{pathFolderCDF}{path to the folder containing the CDF to process.}
}
\value{
A 3 dimensional array containing all the data in the files following the structure: |sample, mz, time2D, time1D|
}
\description{
Read a single CDF file and transform the data into a 3D array.
}
