% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_import.R, R/library_import.R
\name{spectramatrix}
\alias{spectramatrix}
\title{spectramatrix}
\usage{
spectramatrix(lis)

spectramatrix(lis)
}
\arguments{
\item{libs}{list obtained with `getMSP` function}
}
\value{
matrix of spectra

a matrix
}
\description{
Converts library entries to matrix

Convert the spectra list to matrix
}
